exercises 17-23: implement token authentication and authorization, added user-blog schema refs, refactored errors and handlers, introduced custom parser middlewared, fixed existing integration tests and added new ones

+++ controllers/blogs.js
@@ -1,26 +1,42 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const { validate_user, validate_authorization } = require('./validation/blogs_validation')
 
 blogsRouter.get('/', async (request, response) =>
-    response.json(await Blog.find({})))
+    response.json(await Blog.find({}).populate('user', { username: 1, name: 1 })))
 
 blogsRouter.post('/', async (request, response) => {
-    const blog = new Blog(request.body)
-    response.status(201).json(await blog.save())
+    await validate_user(request)
+
+    const user = await User.findById(request.user)
+    const blog = await (new Blog({
+        ...request.body,
+        user: user.id
+    })).save()
+
+    user.blogs = user.blogs.concat(blog.id)
+    await user.save()
+
+    response.status(201).json(blog)
 })
 
 blogsRouter.put('/:id', async (req, res) => {
-    const blog = {
-        title: req.body.title,
-        author: req.body.author,
-        url: req.body.url,
-        likes: req.body.likes
-    }
+    await validate_authorization(req)
 
-    res.json(await Blog.findByIdAndUpdate(req.params.id, blog, { new: true, runValidators: true, context: 'query' }))
+    res.json(await Blog.findByIdAndUpdate(
+        req.params.id, {
+            title: req.body.title,
+            author: req.body.author,
+            url: req.body.url,
+            likes: req.body.likes
+        },
+        { new: true, runValidators: true, context: 'query' }))
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
+    await validate_authorization(req)
+
     await Blog.findByIdAndDelete(req.params.id)
     res.status(204).end()
 })
